SLE-1056: Create basic Cirrus CI pipeline (#3)
This creates a simple Cirrus CI pipeline for building the P2 Repository (Eclipse Update Site) and running the ITs (as far as possible in reasonable time, other tests are disabled) in parallel.

It is kept as simple as possible, not leveraging Artifactory (Repox) for moving artifacts around but just uploading the artifact from the `build` task. There is also no analysis using SonarQube Next / SonarQube Cloud as this would be overkill.